# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateEventDraftDto": {
        "properties": {
          "description": {
            "maxLength": 2000,
            "type": "string"
          },
          "eventType": {
            "enum": [
              "wedding",
              "birthday",
              "offsite",
              "hangout",
              "conference",
              "reunion",
              "fundraiser",
              "other"
            ],
            "example": "wedding",
            "type": "string"
          },
          "scheduledDate": {
            "description": "Provisional event date; can change while planning.",
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "example": "Wedding planning — Tunde & Bola",
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "title",
          "eventType"
        ],
        "type": "object"
      },
...
      "EventAddonRefDto": {
        "properties": {
          "createdAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          },
          "slug": {
            "example": "anonymous_proposals",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "createdAt"
        ],
        "type": "object"
      },
...
      "EventResponseDto": {
        "properties": {
          "addons": {
            "items": {
              "$ref": "#/components/schemas/EventAddonRefDto"
            },
            "type": "array"
          },
          "archivedAt": {
            "example": null,
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          },
          "creatorId": {
            "example": "usr_2N9k8x...",
            "type": "string"
          },
          "currency": {
            "example": "NGN",
            "type": "string"
          },
          "currentUserRole": {
            "description": "The viewing user's role on this event. Populated on single-event reads (GET /:id) and after mutations; absent on list responses.",
            "enum": [
              "admin",
              "contributor",
              "observer"
            ],
            "type": "string"
          },
          "customSubdomain": {
            "example": null,
            "nullable": true,
            "type": "string"
          },
          "description": {
            "example": "Outdoor evening reception, ~120 guests.",
            "nullable": true,
            "type": "string"
          },
          "eventType": {
            "example": "wedding",
            "type": "string"
          },
          "features": {
            "additionalProperties": true,
            "description": "Tier-derived feature bag. Shape: { sharing: {...}, branding: {...}, invitees: {...}, … }.",
            "example": {
              "branding": {
                "customSharePage": false
              },
              "invitees": {
                "customQuestions": 5,
                "maxInvitees": 200
              },
              "sharing": {
                "customSlug": true,
                "emailReminders": true
              }
            },
            "type": "object"
          },
          "id": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "provisionedAt": {
            "example": null,
            "nullable": true,
            "type": "string"
          },
          "publishedAt": {
            "example": null,
            "nullable": true,
            "type": "string"
          },
          "scheduledDate": {
            "description": "YYYY-MM-DD; null when no date is locked yet.",
            "example": "2026-06-24",
            "nullable": true,
            "type": "string"
          },
          "shareSlug": {
            "example": "tunde-bola",
            "nullable": true,
            "type": "string"
          },
          "state": {
            "enum": [
              "draft",
              "planning",
              "published",
              "past",
              "archived"
            ],
            "example": "draft",
            "type": "string"
          },
          "tierSlug": {
            "example": "occasion",
            "nullable": true,
            "type": "string"
          },
          "title": {
            "example": "Tunde & Bola — Wedding",
            "type": "string"
          },
          "updatedAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          }
        },
        "required": [
          "id",
          "creatorId",
          "title",
          "eventType",
          "state",
          "features",
          "currency",
          "createdAt",
          "updatedAt",
          "addons"
        ],
        "type": "object"
      },
...
      "HolidayScanResponseDto": {
        "properties": {
          "costCents": {
            "description": "Cost of this call in USD cents. 0 on `latest()` reads.",
            "example": 3,
            "type": "number"
          },
          "scannedDate": {
            "description": "The scheduledDate the scan was run against. The one-shot enforcement uses this so a date change unlocks a re-scan.",
            "example": "2026-06-24",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/HolidayWarningDto"
            },
            "type": "array"
          }
        },
        "required": [
          "warnings",
          "scannedDate",
          "costCents"
        ],
        "type": "object"
      },
      "HolidayWarningDto": {
        "properties": {
          "detail": {
            "example": "June is peak rainy season — outdoor venues at risk. Consider a covered backup or shift to October.",
            "type": "string"
          },
          "severity": {
            "description": "info = worth-knowing context; warning = likely friction; critical = certain conflict.",
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "example": "warning",
            "type": "string"
          },
          "title": {
            "example": "Rainy season in Lagos",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "title",
          "detail"
        ],
        "type": "object"
      },
...
      "PublishBlockerDto": {
        "properties": {
          "details": {
            "additionalProperties": true,
            "description": "Free-form details about the blocker (e.g. which segment is unlocked).",
            "type": "object"
          },
          "kind": {
            "enum": [
              "state_not_publishable",
              "no_scheduled_date",
              "no_segments",
              "unlocked_segments",
              "odd_voting_count"
            ],
            "example": "unlocked_segments",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "type": "object"
      },
      "PublishEventDto": {
        "properties": {
          "customSlug": {
            "description": "Reserved — custom share slug requires Occasion+ tier (Phase 5c).",
            "maxLength": 64,
            "type": "string"
          }
        },
        "type": "object"
      },
      "PublishReadinessCountsDto": {
        "properties": {
          "lockedSegments": {
            "example": 5,
            "type": "number"
          },
          "planningMembers": {
            "example": 4,
            "type": "number"
          },
          "totalSegments": {
            "example": 6,
            "type": "number"
          }
        },
        "required": [
          "totalSegments",
          "lockedSegments",
          "planningMembers"
        ],
        "type": "object"
      },
      "PublishReadinessResponseDto": {
        "properties": {
          "blockers": {
            "items": {
              "$ref": "#/components/schemas/PublishBlockerDto"
            },
            "type": "array"
          },
          "counts": {
            "$ref": "#/components/schemas/PublishReadinessCountsDto"
          },
          "ready": {
            "example": false,
            "type": "boolean"
          }
        },
        "required": [
          "ready",
          "blockers",
          "counts"
        ],
        "type": "object"
      },
...
      "RenameShareSlugDto": {
        "properties": {
          "slug": {
            "description": "New share slug. Lowercase letters, digits, and single hyphens, 3-64 chars. Reserved words (api, admin, etc.) are rejected.",
            "example": "tundes-30th",
            "maxLength": 64,
            "minLength": 3,
            "pattern": "SHARE_SLUG_REGEX",
            "type": "string"
          }
        },
        "required": [
          "slug"
        ],
        "type": "object"
      },
...
      "UnpublishEventDto": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateEventDraftDto": {
        "properties": {
          "description": {
            "maxLength": 2000,
            "type": "string"
          },
          "eventType": {
            "enum": [
              "wedding",
              "birthday",
              "offsite",
              "hangout",
              "conference",
              "reunion",
              "fundraiser",
              "other"
            ],
            "example": "wedding",
            "type": "string"
          },
          "scheduledDate": {
            "description": "Provisional event date; can change while planning.",
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "example": "Wedding planning — Tunde & Bola",
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "cookie": {
        "in": "cookie",
        "name": "better-auth.session_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Event Planning Decision Making Assistant. Spec frozen per Phase 5.5 — bump `info.version` (major for breaking, minor for additive) on every committed change.",
    "title": "EPDEMA API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/events": {
      "get": {
        "operationId": "events_list",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Events created by the authenticated user or where they are a planning member."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the caller's events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "events_createDraft",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDraftDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            },
            "description": "Create a draft event. No payment yet — tier is chosen later."
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a draft event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/events/{eventId}/holiday-scan": {
      "get": {
        "operationId": "holidayScanner_latest",
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolidayScanResponseDto"
                }
              }
            },
            "description": "Returns the most recent persisted scan (warnings + scannedDate) so the event detail page can surface them without re-rolling Claude. When nothing's been scanned yet, returns `{ warnings: [], scannedDate: '', costCents: 0 }`."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the most recent holiday-scan result (no new model call)",
        "tags": [
          "holiday-scanner"
        ]
      },
      "post": {
        "operationId": "holidayScanner_scan",
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolidayScanResponseDto"
                }
              }
            },
            "description": "Generates a fresh scan against the current event.scheduledDate. Returns 403 when the event doesn't have the `holiday_scanner` add-on; 409 when a scan already exists for this date (re-run after changing the date, or buy the add-on again to re-roll)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run the holiday & conflict scan for an event (admin, add-on required)",
        "tags": [
          "holiday-scanner"
        ]
      }
    },
...
    "/events/{id}": {
      "get": {
        "operationId": "events_get",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            },
            "description": "Get an event the authenticated user is a planning member of (or owns)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one event",
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "events_updateDraft",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDraftDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            },
            "description": "Update a draft event (title, type, date, description)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a draft event",
        "tags": [
          "events"
        ]
      }
    },
    "/events/{id}/mark-past": {
      "post": {
        "operationId": "events_markPast",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            },
            "description": "Phase 7·A — admin-only manual transition from published → past. Required before planning members can submit reviews. Idempotent on already-past events. 400 when state is not published, or when scheduledDate is null/future. In 7·D, a worker will fire this automatically once the scheduled date passes."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Flip a published event to past state (admin)",
        "tags": [
          "events"
        ]
      }
    },
    "/events/{id}/publish": {
      "post": {
        "operationId": "events_publish",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishEventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            },
            "description": "Publish the event (admin only). Validates all blockers; generates a share slug; broadcasts event.changed."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish the event (admin)",
        "tags": [
          "events"
        ]
      }
    },
    "/events/{id}/publish-readiness": {
      "get": {
        "operationId": "events_publishReadiness",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReadinessResponseDto"
                }
              }
            },
            "description": "Which blockers remain (unlocked segments, missing date, odd voting count). Used by the pre-publish modal."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Compute publish-readiness",
        "tags": [
          "events"
        ]
      }
    },
    "/events/{id}/share-slug": {
      "patch": {
        "operationId": "events_renameShareSlug",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameShareSlugDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            },
            "description": "Rename the public share slug (admin only, Occasion-tier feature). Format: lowercase letters/digits/single hyphens, 3-64 chars. Reserved words and taken slugs return 409."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename the public share slug (admin)",
        "tags": [
          "events"
        ]
      }
    },
    "/events/{id}/unpublish": {
      "post": {
        "operationId": "events_unpublish",
        "parameters": [
          {
            "description": "Event UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnpublishEventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            },
            "description": "Roll back to planning state (admin only). Preserves the share slug so re-publishing keeps the same URL."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Roll back to planning (admin)",
        "tags": [
          "events"
        ]
      }
    },
...
  },
  "servers": [],
...
}